Reorganize sovereign-client-components module
