+++ crates/sovereign/sovereign-client-components/src/cosmos/impls/mod.rs
@@ -1,0 +1,1 @@
+pub mod client;
