+++ crates/sovereign/sovereign-client-components/src/cosmos/mod.rs
@@ -3,4 +3,4 @@
    with the counterparty being a Sovereign rollup
 */
 
-pub mod client;
+pub mod impls;
