--- crates/sovereign/sovereign-client-components/src/impls/mod.rs
@@ -1,2 +1,0 @@
-pub mod cosmos;
-pub mod sovereign;
